SmartGrid_Project_Devops/mlflow/models/requirements.txt
mlflow==2.8.1
scikit-learn==1.3.2
pandas==2.1.0
numpy==1.24.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0

